feat: Add react-awesome-slider package for image carousel
Introduce the react-awesome-slider package to the project, enabling the implementation of an image carousel. This package provides a user-friendly and customizable carousel component that enhances the visual experience of the Hero page. The addition of this package allows for the creation of a dynamic and engaging image slideshow on the homepage.